+++ adapters/AdColony/AdColonyAdapter/GADMAdapterAdColony.m
@@ -180,7 +180,7 @@ typedef enum {
 @implementation GADMAdapterAdColony
 
 + (NSString *)adapterVersion {
-  return @"3.1.1.1";
+  return @"3.2.0.0";
 }
 
 + (Class<GADAdNetworkExtras>)networkExtrasClass {