+++ adapters/Unity/UnityAdapter/GADMAdapterUnityConstants.h
@@ -23,7 +23,7 @@ static NSString *const GADMAdapterUnityGameID = @"gameId";
 static NSString *const GADMAdapterUnityPlacementID = @"zoneId";
 
 /// Ad mediation network adapter version.
-static NSString *const GADMAdapterUnityVersion = @"2.1.0.0";
+static NSString *const GADMAdapterUnityVersion = @"2.1.1.0";
 
 /// Ad mediation network name.
 static NSString *const GADMAdapterUnityMediationNetworkName = @"AdMob";