+++ example/Adapter/SampleAdapter.m
@@ -16,13 +16,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
+@import SampleAdSDK;
 
 #import "SampleAdapter.h"
-
-#import "../SDK/SampleBanner.h"
-#import "../SDK/SampleInterstitial.h"
-#import "../SDK/SampleNativeAdLoader.h"
-#import "../SDK/SampleRewardBasedVideo.h"
 #import "SampleAdapterDelegate.h"
 #import "SampleAdapterMediatedNativeAppInstallAd.h"
 #import "SampleAdapterMediatedNativeContentAd.h"