+++ example/CustomEvent/SampleMediatedNativeAppInstallAd.m
@@ -19,7 +19,6 @@
 
 #import "SampleMediatedNativeAppInstallAd.h"
 
-#import "../SDK/SampleAdInfoView.h"
 #import "SampleCustomEventConstants.h"
 
 // You may notice that this class and the Mediation Adapter's