+++ example/SampleAdSDK/SampleAdSDKAdapter/SampleAdSDKAdapter.h
@@ -1,0 +1,23 @@
+//
+//  SampleAdSDKAdapter.h
+//  SampleAdSDKAdapter
+//
+//  Created by Samuel Stow on 9/6/17.
+//  Copyright © 2017 Google. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SampleAdSDKAdapter.
+FOUNDATION_EXPORT double SampleAdSDKAdapterVersionNumber;
+
+//! Project version string for SampleAdSDKAdapter.
+FOUNDATION_EXPORT const unsigned char SampleAdSDKAdapterVersionString[];
+
+#import <SampleAdSDKAdapter/SampleAdapter.h>
+#import <SampleAdSDKAdapter/SampleAdapterDelegate.h>
+#import <SampleAdSDKAdapter/SampleAdapterConstants.h>
+#import <SampleAdSDKAdapter/SampleAdapterProtocol.h>
+#import <SampleAdSDKAdapter/SampleAdapterMediatedNativeAppInstallAd.h>
+#import <SampleAdSDKAdapter/SampleAdapterMediatedNativeContentAd.h>
+