+++ adapters/AdColony/CHANGELOG.md
@@ -1,5 +1,8 @@
 # AdColony Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Version 3.2.0.0
+- Verified compatibility with AdColony SDK 3.2.0.
+
 ## 3.1.1.1
 - Removed support for the `armv7s` architecture.
 - Fixed an issue that caused the adapter to incorrectly invoke the rewarded