+++ adapters/Chartboost/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Chartboost Ads Mediation Adapter for Google Mobile Ads SDK for iOS
 
+## Version 7.0.0.0
+- Verified compatibility with Chartboost SDK 7.0.0.
+
 ## Version 6.6.3.0
 - Verified compatibility with Chartboost SDK 6.6.3.
 - Removed the support for 'armv7s' architecture.